feat(区域化翻译): ✨ 从模块导入翻译接口，并添加123个语言代码
